Chromium 38.0.2125.44 mac beta

+++ src/tools/telemetry/telemetry/core/backends/chrome/desktop_browser_backend.py
@@ -2,7 +2,6 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-import distutils
 import glob
 import heapq
 import logging
@@ -18,6 +17,7 @@ from telemetry.core import exceptions
 from telemetry.core import util
 from telemetry.core.backends import browser_backend
 from telemetry.core.backends.chrome import chrome_browser_backend
+from telemetry.util import path
 from telemetry.util import support_binaries
 
 
@@ -28,7 +28,7 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
   def __init__(self, browser_options, executable, flash_path, is_content_shell,
                browser_directory, output_profile_path, extensions_to_load):
     super(DesktopBrowserBackend, self).__init__(
-        is_content_shell=is_content_shell,
+        supports_tab_control=not is_content_shell,
         supports_extensions=not is_content_shell,
         browser_options=browser_options,
         output_profile_path=output_profile_path,
@@ -45,6 +45,8 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
 
     assert not flash_path or os.path.exists(flash_path)
     self._flash_path = flash_path
+
+    self._is_content_shell = is_content_shell
 
     if len(extensions_to_load) > 0 and is_content_shell:
       raise browser_backend.ExtensionsNotSupportedException(
@@ -69,7 +71,7 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
         self._tmp_profile_dir = tempfile.mkdtemp()
       profile_dir = self._profile_dir or self.browser_options.profile_dir
       if profile_dir:
-        if self.is_content_shell:
+        if self._is_content_shell:
           logging.critical('Profiles cannot be used with content shell')
           sys.exit(1)
         logging.info("Using profile directory:'%s'." % profile_dir)
@@ -105,11 +107,7 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
         '--pipe-name=%s' % self._GetCrashServicePipeName()])
 
   def _GetCdbPath(self):
-    search_paths = [os.getenv('PROGRAMFILES(X86)', ''),
-                    os.getenv('PROGRAMFILES', ''),
-                    os.getenv('LOCALAPPDATA', ''),
-                    os.getenv('PATH', '')]
-    possible_paths = [
+    possible_paths = (
         'Debugging Tools For Windows',
         'Debugging Tools For Windows (x86)',
         'Debugging Tools For Windows (x64)',
@@ -119,13 +117,12 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
                      'x86'),
         os.path.join('win_toolchain', 'vs2013_files', 'win8sdk', 'Debuggers',
                      'x64'),
-        ]
+    )
     for possible_path in possible_paths:
-      path = distutils.spawn.find_executable(
-          os.path.join(possible_path, 'cdb'),
-          path=os.pathsep.join(search_paths))
-      if path:
-        return path
+      app_path = os.path.join(possible_path, 'cdb.exe')
+      app_path = path.FindInstalledWindowsApplication(app_path)
+      if app_path:
+        return app_path
     return None
 
   def HasBrowserFinishedLaunching(self):
@@ -169,7 +166,7 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
     args.append('--remote-debugging-port=%i' % self._port)
     args.append('--enable-crash-reporter-for-testing')
     args.append('--use-mock-keychain')
-    if not self.is_content_shell:
+    if not self._is_content_shell:
       args.append('--window-size=1280,1024')
       if self._flash_path:
         args.append('--ppapi-flash-path=%s' % self._flash_path)
@@ -181,7 +178,7 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
     # Make sure _profile_dir hasn't already been set.
     assert self._profile_dir is None
 
-    if self.is_content_shell:
+    if self._is_content_shell:
       logging.critical('Profile creation cannot be used with content shell')
       sys.exit(1)
 
@@ -209,7 +206,6 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
 
     try:
       self._WaitForBrowserToComeUp()
-      self._PostBrowserStartupInitialization()
     except:
       self.Close()
       raise
@@ -321,8 +317,8 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
   def Close(self):
     super(DesktopBrowserBackend, self).Close()
 
-    # First, try to politely shutdown.
-    if self.IsBrowserRunning():
+    # Shutdown politely if the profile may be used again.
+    if self._output_profile_path and self.IsBrowserRunning():
       self._proc.terminate()
       try:
         util.WaitFor(lambda: not self.IsBrowserRunning(), timeout=5)
@@ -330,7 +326,7 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
       except util.TimeoutException:
         logging.warning('Failed to gracefully shutdown. Proceeding to kill.')
 
-    # If it didn't comply, get more aggressive.
+    # Shutdown aggressively if the above failed or if the profile is temporary.
     if self.IsBrowserRunning():
       self._proc.kill()
